marathon-mgmt/tonomi-files/requirements.txt
unittest
requests_mock
pyyaml